Incorporate the ANSI helper into helper file
